Limit worker scheduling by occupancy
The worker scheduler would previously assume that it was fine to
schedule infinite amounts of work onto the same processor at once, which
is only efficient when tasks do lots of `yield`ing.

This commit teaches `@spawn` and the worker scheduler about a new
`occupancy` task option, which is a value between 0 and 1 which
approximates how fully the task occupies the processor. If the occupancy
is 0.2, then 5 such tasks can execute concurrently and fully occupy the
processor.

Processors now operate primarily from a single controlling task per
processor, and work queues up in a lowest-occupancy-first manner.